requirements.txt
streamlit
google-generativeai
deep-translator
opencage
requests
Pillow
requests
python-dotenv
flask
flask-wtf